Move code to new common.ModelDataWithVelocityRepresentation class
